+++ server/types/price-action.types.ts
@@ -1,23 +1,5 @@
 import type { OHLCV } from "@/lib/polygon/polygon.types";
-import { datelike, numeric } from "@shared/types/zod.utility.types";
-import { z } from "zod";
-
-export const priceActionSchema = z.object({
-	ticker: z.string(),
-	timestamp: datelike, // unix ms timestamp, or date, depending on whether we parse inside the query
-	open: numeric(6),
-	close: numeric(6),
-	high: numeric(6),
-	low: numeric(6),
-	volume: numeric(2),
-});
-
-export const priceActionWithUpdatedAtSchema = priceActionSchema.extend({
-	updated_at: datelike, // same note as with `timestamp` in `priceActionSchema`
-});
-
-export type PriceAction = z.infer<typeof priceActionSchema>;
-export type PriceActionWithUpdatedAt = z.infer<typeof priceActionWithUpdatedAtSchema>;
+import type { PriceAction } from "@shared/types/price-action.types";
 
 const polygonToPriceAction = {
 	ticker: "T",
